+++ spec/zenlish/lang/dictionary_spec.rb
@@ -27,16 +27,21 @@ module Zenlish
           expect(common_nouns.all? { |c_n| lemmas.include?(c_n) }).to be_truthy
         end
 
-        let(:present_1sg) { [:first, :singular, :present, nil] }
-        let(:present_3sg) { [:third, :singular, :present, nil] }
-        let(:present_1pl) { [:first, :plural,   :present, nil] }
-        let(:progressive) { [nil,    nil,       :progressive, nil] }
-        let(:past_simple) { [nil,    nil,       :past_simple, nil] }
-        let(:past_participle) { [nil, nil,      :past_participle, nil] }
+        let(:present_1sg)     { [:first,  :singular, :present, nil] }
+        let(:present_2sg)     { [:second, :singular, :present, nil] }        
+        let(:present_3sg)     { [:third,  :singular, :present, nil] }
+        let(:present_1pl)     { [:first,  :plural,   :present, nil] }
+        let(:progressive)     { [nil,     nil,       :progressive, nil] }
+        let(:past_simple)     { [:first,  :singular, :past_simple, nil] }
+        let(:past_simple_2sg) { [:second, :singular, :past_simple, nil] }
+        let(:past_simple_3sg) { [:third,  :singular, :past_simple, nil] }
+        let(:past_simple_1pl) { [:first,  :plural,   :past_simple, nil] }         
+        let(:past_participle) { [nil,     nil,       :past_participle, nil] }
 
-        def test_inflection_of(verb_lexeme, pairs)
+        def test_inflection_of(verb_lexeme, pairs, drop_last_column = false)
           pairs.each do |(constraints, expected_form)|
-            expect(verb_lexeme.inflect(constraints)).to eq(expected_form)
+            restriction = drop_last_column ? constraints.slice(0..-2) : constraints
+            expect(verb_lexeme.inflect(restriction)).to eq(expected_form)
           end
         end
 
@@ -61,8 +66,37 @@ module Zenlish
             [past_simple,     'did'],
             [past_participle, 'done']
           ]
-          test_inflection_of(lexm, expectations_2)          
+          test_inflection_of(lexm, expectations_2, true)
+
+          lexm = subject.get_lexeme('have', WClasses::IrregularVerbHave)
+          expectations_3 = [
+            [present_1sg,     'have'],
+            [present_3sg,     'has'],
+            [present_1pl,     'have'],
+            [progressive,     'having'],
+            [past_simple,     'had'],
+            [past_participle, 'had']
+          ]
+          test_inflection_of(lexm, expectations_3, true)           
         end
+        
+        it 'should know how to inflect the verb be' do
+          lexm = subject.get_lexeme('be', WClasses::IrregularVerbBe)
+          expectations_1 = [
+            [present_1sg,     'am'],
+            [present_2sg,     'are'],
+            [present_3sg,     'is'],
+            [present_1pl,     'are'],
+            [progressive,     'being'],
+            [past_simple,     'was'],
+            [past_simple_2sg, 'were'],
+            [past_simple_3sg, 'was'],
+            [past_simple_1pl, 'were'],
+            [past_participle, 'been']           
+          ]
+          test_inflection_of(lexm, expectations_1, true)        
+        end
+        
       end # context
     end # describe
   end # module